+++ backend/app/schemas/career_preference.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict
@@ -8,14 +7,14 @@ class CareerPreferenceBase(BaseModel):
     target_roles: list[str] | None = None
     location_preference: str | None = None
     work_mode: str | None = None
-    salary_minimum: Decimal | None = None
+    salary_minimum: int | None = None
 
 
 class CareerPreferenceUpdate(BaseModel):
     target_roles: list[str] | None = None
     location_preference: str | None = None
     work_mode: str | None = None
-    salary_minimum: Decimal | None = None
+    salary_minimum: int | None = None
 
 
 class CareerPreferenceRead(CareerPreferenceBase):
